rework file structure, rename
