+++ rotala-client/src/strategy/staticweight.rs
