+++ rotala-http/src/bin/uist_server_v1.rs
@@ -2,12 +2,11 @@ use std::env;
 use std::sync::Mutex;
 
 use actix_web::{web, App, HttpServer};
-use rotala::{
-    http::uist_v1::{
-        server::{delete_order, fetch_quotes, info, init, insert_order, tick},
-        AppState,
-    },
-    input::penelope::Penelope,
+
+use rotala::input::penelope::Penelope;
+use rotala_http::http::uist_v1::{
+    server::{delete_order, fetch_quotes, info, init, insert_order, tick},
+    AppState,
 };
 
 #[actix_web::main]
