+++ rotala-http/src/bin/uist_server_v2.rs
@@ -2,13 +2,10 @@ use std::env;
 use std::sync::Mutex;
 
 use actix_web::{web, App, HttpServer};
-use rotala::{
-    http::jura_v1::{
-        jurav1_server::{delete_order, fetch_quotes, info, init, insert_order, tick},
-        AppState,
-    },
-    input::penelope::Penelope,
-};
+use rotala::input::athena::Athena;
+use rotala_http::http::uist_v2::server::*;
+use rotala_http::http::uist_v2::AppState;
+use std::path::Path;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
@@ -16,21 +13,21 @@ async fn main() -> std::io::Result<()> {
 
     let address: String = args[1].clone();
     let port: u16 = args[2].parse().unwrap();
+    let file_path = Path::new(&args[3]);
 
-    let source = Penelope::random(3000, vec!["0"]);
-    let app_state = AppState::single("RANDOM", source);
+    let source = Athena::from_file(file_path);
+    let app_state = AppState::single("Test", source);
 
-    let jura_state = web::Data::new(Mutex::new(app_state));
+    let uist_state = web::Data::new(Mutex::new(app_state));
 
     HttpServer::new(move || {
         App::new()
-            .app_data(jura_state.clone())
+            .app_data(uist_state.clone())
             .service(info)
             .service(init)
             .service(fetch_quotes)
             .service(tick)
             .service(insert_order)
-            .service(delete_order)
     })
     .bind((address, port))?
     .run()
