+++ rotala-http/src/http/uist_v2.rs
@@ -5,8 +5,8 @@ use std::sync::Mutex;
 use anyhow::{Error, Result};
 use serde::{Deserialize, Serialize};
 
-use crate::exchange::uist_v2::{Order, Trade, UistV2};
-use crate::input::athena::{Athena, DateBBO};
+use rotala::exchange::uist_v2::{Order, Trade, UistV2};
+use rotala::input::athena::{Athena, DateBBO};
 
 type BacktestId = u64;
 
@@ -504,9 +504,9 @@ pub mod server {
 mod tests {
     use actix_web::{test, web, App};
 
-    use crate::exchange::uist_v2::Order;
     use crate::http::uist_v1::NowResponse;
-    use crate::input::athena::Athena;
+    use rotala::exchange::uist_v2::Order;
+    use rotala::input::athena::Athena;
 
     use super::server::*;
     use super::{AppState, FetchQuotesResponse, InitResponse, InsertOrderRequest, TickResponse};
